# api/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Cinnabar API
  description: >-
    This is a Movies/Tv-series Reviews Server based on the OpenAPI 3.0
    specification. 
     If you want to know more about the project here are some useful links:
    - [The Tooren Romaros
    repository](https://github.com/CoPhos/ToorenRomaros/tree/master)

    - [The source API definition for the Tooren Romaros
    project](https://github.com/CoPhos/ToorenRomaros/blob/develop/tooren_romaros_api/src/main/resources/api/openapi.yaml)
  contact:
    name: Jorge Cobos
    url: https://www.JorgeCobos.com
    email: cobos1932@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: "1.0"
servers:
  - url: http://localhost:9090
    description: Development
  - url: http://localhost:9090
    description: Production
...
paths:
...
  /api/v1/staffs:
    post:
      tags:
        - post
        - Staff
        - Saga
      summary: Create Stafff
      description: >-
        This endpoint allows users to create a new staff by providing the
        necessary details using the provided payload.
      operationId: createStaff
      parameters:
        - name: GetStaffDto
          in: query
          description: GetStaffDto object
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStaffDto"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffDto"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateStaffDto"
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
...
  /api/v1/films/staffs:
    post:
      tags:
        - post
        - Film
        - Staff
      summary: Create Film-Staff Relationship
      description: >-
        This endpoint allows users to create a new relationship between a film
        and a staff member by providing the necessary details using the provided
        payload.
      operationId: createStaffFilm
      parameters:
        - name: CreateStaffFilmDto
          in: query
          description: CreateStaffFilmDto object
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateStaffFilmDto"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffFilmDto"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStaffFilmDto"
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
...
  /api/v1/staffs/{staffId}:
    get:
      tags:
        - Staff
        - get
      summary: Get Staff by ID
      description: >-
        This endpoint retrieves a specific staff identified by its unique
        identifier (id).
      operationId: getStaff
      parameters:
        - name: staffId
          in: path
          description: ID of the Staff
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStaffDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
    delete:
      tags:
        - Staff
        - delete
      summary: Delete Staff
      description: >-
        This endpoint allows users to delete a specific staff identified by its
        unique identifier (id). Deleting a staff is a permanent action and
        cannot be undone.
      operationId: deleteStaff
      parameters:
        - name: staffId
          in: path
          description: The ID of the Staff
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "202":
          description: Accepted
          content:
            text/plain: {}
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
    patch:
      tags:
        - patch
        - Staff
      summary: Update Staff
      description: >-
        This endpoint allows users to update a specific Staff entity identified
        by its unique identifier (staffId). Users can modify the properties of
        the staff using the provided update payload.
      operationId: updateStaff
      parameters:
        - name: staffId
          in: path
          description: ID of the Staff
          required: true
        - name: UpdateStaffDto
          in: query
          description: UpdateStaffDto object
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateStaffDto"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStaffDto"
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - Bearer Authentication:
            - ADMIN
...
  /api/v1/films/{filmId}/staffs:
    get:
      tags:
        - Film
        - Staff
        - get
      summary: Get All Staff From a Specific Film
      description: >-
        This endpoint retrieves all entities of the film-Staff relationship for
        a specific film identified by its unique identifier (filmId).
      operationId: getStaffFilmByFilmId
      parameters:
        - name: filmId
          in: path
          description: ID of the Film
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStaffFilmDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
...
  /api/v1/films/staffs/{relationshipId}:
    delete:
      tags:
        - Film
        - Staff
        - delete
      summary: Delete Film-Staff Relationship
      description: >-
        This endpoint allows users to delete a specific film-Staff relationship
        identified by its unique identifier (relationshipId). Deleting a
        relationship is a permanent action and cannot be undone.
      operationId: deleteStaffFilm
      parameters:
        - name: relationshipId
          in: path
          description: The ID of the Relationship
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "202":
          description: Accepted
          content:
            text/plain: {}
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
components:
  schemas:
...
    Error:
      type: object
      properties:
        errorCode:
          type: string
        message:
          type: string
        status:
          type: integer
          format: int32
        url:
          type: string
        reqMethod:
          type: string
        timestamp:
          type: string
          format: date-time
...
    CreateStaffDto:
      required:
        - biography
        - birthday
        - birthplace
        - firstName
        - lastName
      type: object
      properties:
        firstName:
          maxLength: 50
          minLength: 0
          type: string
        lastName:
          maxLength: 50
          minLength: 0
          type: string
        biography:
          type: string
        birthday:
          type: string
          format: date
        birthplace:
          maxLength: 50
          minLength: 0
          type: string
    GetStaffDto:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        biography:
          type: string
        birthday:
          type: string
          format: date
        birthplace:
          type: string
...
    CreateStaffFilmDto:
      required:
        - filmName
        - role
        - staffId
      type: object
      properties:
        role:
          maxLength: 50
          minLength: 0
          type: string
        appearAs:
          maxLength: 50
          minLength: 0
          type: string
        filmName:
          type: string
        staffId:
          type: string
    GetStaffFilmDto:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
        appearAs:
          type: string
        filmName:
          type: string
        staffName:
          type: string
        staffId:
          type: string
...
    UpdateStaffDto:
      type: object
      properties:
        firstName:
          maxLength: 50
          minLength: 0
          type: string
        lastName:
          maxLength: 50
          minLength: 0
          type: string
        biography:
          type: string
        birthday:
          type: string
          format: date
        birthplace:
          maxLength: 50
          minLength: 0
          type: string
...
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT